recover from failed proposals (#116)

+++ components/app/claim/stake/editor/StakeButtonsOverlay.tsx
@@ -19,6 +19,7 @@ import { XMarkIcon } from "@/components/icon/XMarkIcon";
 interface Props {
   claim: ClaimObject;
   open: string;
+  pending: boolean;
   setOpen: (open: string) => void;
 }
 
@@ -29,22 +30,54 @@ export const StakeButtonsOverlay = (props: Props) => {
   const editor = EditorStore.getState();
 
   React.useEffect(() => {
-    if (props.open === "") {
+    if (props.open === "" && !props.pending) {
       editor.delete();
     } else {
       editor.updateClaim(props.claim.id())
       editor.updateContract(props.claim.contract())
-      editor.updateMinimum(props.claim.summary().post.minimum)
       editor.updateToken(props.claim.token())
+
+      // During normal staking, the user decides what side to stake on in
+      // advance. If the user for whatever reason fails to propose their claim
+      // properly, then we have to provide them with an alternative path to
+      // finish their original intention. At the moment we are always setting
+      // the side to true when proposing a claim. And so this is what we have to
+      // do here as well.
+      if (props.pending) {
+        const vot = props.claim.getVote();
+
+        if (vot.length === 1) {
+          editor.updateMinimum(parseFloat(vot[0].value))
+          editor.updateValue(vot[0].value);
+        }
+
+        editor.updateOption("true");
+      } else {
+        editor.updateMinimum(props.claim.summary().post.minimum)
+      }
     }
-  }, [props.claim, props.open, editor]);
+  }, [props.claim, props.open, props.pending, editor]);
 
   return (
     <div className="absolute bottom-12 w-full background">
-      <div className="relative flex mb-2 p-4 bg-sky-100 dark:bg-sky-900 rounded">
+      <div
+        className={`
+          relative flex mb-2 p-4 rounded
+          ${props.pending ? "bg-rose-200 dark:bg-rose-900" : "bg-sky-200 dark:bg-sky-900"}
+        `}
+      >
         <div className="w-full text-sm">
-          You are staking to <strong>{props.open}</strong> with this claim.
-          There is no guarantee of repayment. Learn more at&nbsp;
+          {props.pending ? (
+            <>
+              Your claim could <strong>not</strong> be confirmed onchain.
+              Please try to finish your proposal again. Learn more at&nbsp;
+            </>
+          ) : (
+            <>
+              You are staking to <strong>{props.open}</strong> with this claim.
+              There is no guarantee of repayment. Learn more at&nbsp;
+            </>
+          )}
 
           <Link
             className="text-blue-600 dark:text-blue-400"
@@ -60,6 +93,7 @@ export const StakeButtonsOverlay = (props: Props) => {
         <div className="absolute top-0 right-0">
           <BaseButton
             background="none"
+            color="text-gray-500 dark:text-gray-400"
             icon={<XMarkIcon />}
             onClick={() => props.setOpen("")}
           />
@@ -69,6 +103,7 @@ export const StakeButtonsOverlay = (props: Props) => {
       <div className="flex gap-x-2 h-14">
         <div className="flex-1 w-full">
           <ValueField
+            pending={props.pending}
             setOpen={props.setOpen}
             summary={props.claim.summary()}
             token={props.claim.token()}
@@ -85,6 +120,7 @@ export const StakeButtonsOverlay = (props: Props) => {
             type="button"
             onClick={() => {
               SubmitForm({
+                pending: props.pending ? props.claim : undefined,
                 after: () => {
                   setProcessing("Confirming Onchain");
                 },
@@ -125,7 +161,15 @@ export const StakeButtonsOverlay = (props: Props) => {
                 </div>
               ) : (
                 <div>
-                  Stake to {ToTitle(props.open)}
+                  {props.pending ? (
+                    <div>
+                      Propose Claim
+                    </div>
+                  ) : (
+                    <div>
+                      Stake to {ToTitle(props.open)}
+                    </div>
+                  )}
                 </div>
               )}
             </>
